API surface has been drastically simplified, still retaining its functions
