+++ cryptipass_test.go
@@ -23,6 +23,77 @@ func TestBasic(t *testing.T) {
 	}
 }
 
+type CertifyResult struct {
+	NominalH float64
+	Gap      float64
+	StdDev   float64
+}
+
+// Certify evaluates the entropy and randomness of passphrases generated by a given function.
+//
+// The Certify function runs a comprehensive statistical analysis on the provided password generator
+// function `Gen` by simulating trials of passphrase generation. It computes the average entropy
+// and monitors the gap between the expected entropy and the actual entropy based on the frequency
+// distribution of passphrases generated.
+//
+// It returns a CertifyResult struct, which includes:
+// - NominalH: The nominal entropy of the passphrases, averaged over all trials.
+// - Gap: The difference between the nominal entropy and the actual observed entropy.
+// - StdDev: The standard deviation of the nominal entropy across trials, giving a measure of variability.
+//
+// This function is useful for verifying the strength and unpredictability of passphrases generated by
+// custom implementations of password generators.
+//
+// Parameters:
+// - Gen: A function that generates a passphrase and returns it alongside its entropy.
+//
+// Returns:
+//
+// - CertifyResult: A struct containing the analysis of the generator's entropy.
+//
+// This process continues until the gap between nominal and actual entropy is small enough,
+// or a sufficient number of trials has been conducted.
+func Certify(Gen func() (string, float64)) CertifyResult {
+	nominal_H := 0.0
+	nominal_H2 := 0.0
+	cnt_nom_H := 0.0
+	for range 1000 {
+		_, nh := Gen()
+		nominal_H += nh
+		nominal_H2 += nh * nh
+		cnt_nom_H++
+	}
+	cnt := make(map[string]int)
+	n := float64(0)
+	Q := 64
+	for {
+		for range Q {
+			w, nh := Gen()
+			nominal_H += nh
+			nominal_H2 += nh * nh
+			cnt_nom_H++
+			cnt[w]++
+			n++
+		}
+		Q += Q / 16
+		m := float64(len(cnt))
+		H := 0.0
+		for _, iC := range cnt {
+			c := float64(iC)
+			p := (c / n)
+			H -= p * math.Log2(p)
+		}
+		H += (m - 1) / (2 * n)
+		nomH := nominal_H / cnt_nom_H
+		nomH2 := nominal_H2 / cnt_nom_H
+		stddev := math.Sqrt(max(nomH2-nomH*nomH, 1e-16))
+		gap := nomH - H
+		if math.Abs(gap) < 0.05 || math.Log2(n) > 3*nomH {
+			return CertifyResult{NominalH: nomH, Gap: gap, StdDev: stddev}
+		}
+	}
+
+}
 func TestCert(t *testing.T) {
 	g := cryptipass.NewInstance()
 	g.Rng = rand.New(rand.NewPCG(37512033, 27996124))
@@ -52,28 +123,27 @@ func TestCert(t *testing.T) {
 			return s, h
 		},
 		func() (string, float64) {
-			r, h := g.PickLength()
+			r, h := g.GenWordLength()
 			return fmt.Sprint(r), h
 		},
 		func() (string, float64) {
 			words := []string{"", "ciao", "gi", "ok", "s"}
 			N := len(words)
 			r := g.Rng.IntN(N)
-			nex, h := g.PickNext(words[r])
+			nex, h := g.GenNextToken(words[r])
 			h += math.Log2(float64(N))
 			return words[r] + nex, h
 		},
 	}
 
 	for i, x := range funcs {
-		X := cryptipass.Certify(x)
+		X := Certify(x)
 		if math.Abs(X.Gap) >= 0.1 {
 			t.Fatal("failed certification of function", i+1, X)
 		} else {
 			t.Log("passed certification of function", i+1)
 		}
 	}
-
 }
 
 func TestGenPassphrase(t *testing.T) {
@@ -94,7 +164,7 @@ func TestGenPassphrase(t *testing.T) {
 func TestPickNext(t *testing.T) {
 	g := cryptipass.NewInstance()
 	seed := "te"
-	next, entropy := g.PickNext(seed)
+	next, entropy := g.GenNextToken(seed)
 	if len(next) != 1 {
 		t.Errorf("Expected string extension to be 1 rune long")
 	}
@@ -106,7 +176,7 @@ func TestPickNext(t *testing.T) {
 
 func TestPickLength(t *testing.T) {
 	g := cryptipass.NewInstance()
-	length, entropy := g.PickLength()
+	length, entropy := g.GenWordLength()
 
 	if length < 3 || length > 9 {
 		t.Errorf("Expected length to be between 3 and 9, got %d", length)
